Clear gem paths after installing ruby-shadow
When ruby-shadow gets installed, chef-client will not have the gem in
the gem paths, so we need to clear the paths to make it work. The test
to detect this case is easy: we simply try to load the gem.

This will fix chef-client daemon so that they will work fine after
ruby-shadow has been installed.