Import authkey as attribute if file already existed
On the founder node, when we generate the authkey, we actually only
import the authkey as an attribute if we create the authkey file. But if
the authkey file already exists, we don't do anything.

The result is that non-founder nodes are stuck because they don't have
access to the authkey.

This can happen after deactivating a pacemaker proposal, if the
attributes have been removed by the user, and then reactivating it.